replace old inner mempool with new
