+++ madara/crates/client/devnet/build.rs
@@ -1,3 +1,3 @@
 fn main() {
-    // build_version::get_or_compile_artifacts(4).expect("Failed to load artifacts");
+    build_version::get_or_compile_artifacts(4).expect("Failed to load artifacts");
 }
