--- madara/crates/client/mempool/proptest-regressions/inner/proptest.txt
@@ -1,8 +1,0 @@
-# Seeds for failure cases proptest has generated in the past. It is
-# automatically read and these particular cases re-run before any
-# novel cases are generated.
-#
-# It is recommended to check this file in to source control so that
-# everyone who runs the test benefits from these saved cases.
-cc e0fe24c0917216b42d3ab91ff4bad7efc965970cc8928b3be2fe78ad1b456a41 # shrinks to (initial_state, transitions, seen_counter) = (MempoolInner { nonce_mapping: {}, tx_intent_queue_ready: {}, tx_intent_queue_pending_by_nonce: {}, tx_intent_queue_pending_by_timestamp: {}, deployed_contracts: DeployedContracts({}), limiter: MempoolLimiter { config: MempoolLimits { max_transactions: 18446744073709551615, max_declare_transactions: 18446744073709551615, max_age: Some(3600s) }, current_transactions: 0, current_declare_transactions: 0 } }, [Pop, Push { tx: MempoolTransaction { tx_hash: 0x0, nonce: 0x0, contract_address: 0x2b, tx_type: InvokeFunction, arrived_at: SystemTime { tv_sec: 1736763022, tv_nsec: 105722909 } }, force: false, nonce_info: NonceInfo { readiness: Ready, nonce: Nonce(0x0), nonce_next: Nonce(0x1) } }, Push { tx: MempoolTransaction { tx_hash: 0x0, nonce: 0x0, contract_address: 0x2b, tx_type: InvokeFunction, arrived_at: SystemTime { tv_sec: 1736763786, tv_nsec: 335204314 } }, force: false, nonce_info: NonceInfo { readiness: Pending, nonce: Nonce(0x1), nonce_next: Nonce(0x2) } }, Pop, Push { tx: MempoolTransaction { tx_hash: 0x0, nonce: 0x0, contract_address: 0x2b, tx_type: InvokeFunction, arrived_at: SystemTime { tv_sec: 1736765304, tv_nsec: 335666467 } }, force: true, nonce_info: NonceInfo { readiness: Ready, nonce: Nonce(0x0), nonce_next: Nonce(0x1) } }, Pop, Pop], None)
-cc 50c3ac55269d0e60677604a18a7c72064a812f7ea5f8e15d8ca6b5d36e7520a6 # shrinks to (initial_state, transitions, seen_counter) = (MempoolInner { nonce_mapping: {}, tx_intent_queue_ready: {}, tx_intent_queue_pending_by_nonce: {}, tx_intent_queue_pending_by_timestamp: {}, deployed_contracts: DeployedContracts({}), limiter: MempoolLimiter { config: MempoolLimits { max_transactions: 18446744073709551615, max_declare_transactions: 18446744073709551615, max_age: Some(3600s) }, current_transactions: 0, current_declare_transactions: 0 }, nonce_by_account: {} }, [Push { tx: MempoolTransaction { tx_hash: 0x0, nonce: 0x0, nonce_next: 0x1, contract_address: 0x38, tx_type: InvokeFunction, arrived_at: SystemTime { tv_sec: 1736763594, tv_nsec: 612309376 } }, force: false, nonce_info: NonceInfo { readiness: Ready, nonce: Nonce(0x0), nonce_next: Nonce(0x1) } }], None)
